move the draft definitions to their own file
