+++ libs/newsletters-data-client/src/lib/newsletter-data-type.ts
@@ -178,27 +178,3 @@ export function isPartialNewsletterData(
 ): subject is Partial<NewsletterData> {
 	return newsletterDataSchema.partial().safeParse(subject).success;
 }
-
-export const draftNewsletterDataSchema = newsletterDataSchema.deepPartial();
-export type DraftNewsletterData = z.infer<typeof draftNewsletterDataSchema>;
-
-export function isDraftNewsletterData(
-	subject: unknown,
-): subject is DraftNewsletterData {
-	return draftNewsletterDataSchema.safeParse(subject).success;
-}
-
-export type DraftNewsletterDataWithMeta = DraftNewsletterData & {
-	meta: MetaData;
-};
-export type DraftNewsletterDataWithoutMeta = DraftNewsletterData & {
-	meta: undefined;
-};
-
-export function isDraftNewsletterDataWithMeta(
-	subject: unknown,
-): subject is DraftNewsletterDataWithMeta {
-	return draftNewsletterDataSchema
-		.extend({ meta: metaDataSchema })
-		.safeParse(subject).success;
-}
