+++ libs/newsletters-data-client/src/lib/newsletter-storage/InMemoryNewsletterStorage.ts
@@ -1,6 +1,6 @@
+import type { DraftNewsletterDataWithMeta } from '../draft-newsletter-data-type';
 import { makeBlankMeta } from '../meta-data-type';
 import type {
-	DraftNewsletterDataWithMeta,
 	NewsletterData,
 	NewsletterDataWithMeta,
 	NewsletterDataWithoutMeta,
