+++ libs/newsletters-data-client/src/lib/newsletter-storage/NewsletterStorage.ts
@@ -1,12 +1,12 @@
+import type { DraftNewsletterDataWithMeta } from '../draft-newsletter-data-type';
 import type { MetaData } from '../meta-data-type';
 import { createNewMeta, stripMeta, updateMeta } from '../meta-data-type';
-import { isPartialNewsletterData } from '../newsletter-data-type';
 import type {
-	DraftNewsletterDataWithMeta,
 	NewsletterData,
 	NewsletterDataWithMeta,
 	NewsletterDataWithoutMeta,
 } from '../newsletter-data-type';
+import { isPartialNewsletterData } from '../newsletter-data-type';
 import type {
 	SuccessfulStorageResponse,
 	UnsuccessfulStorageResponse,
